Routing as behavior
