+++ lib/nav_view/nav_view.html
@@ -23,6 +23,7 @@ Left Nav + View
                 <!-- Main Toolbar -->
                 <paper-toolbar>
                     <paper-icon-button icon="menu" paper-drawer-toggle></paper-icon-button>
+                    <div class="flex">{{selectedPage.name}}</div>
                 </paper-toolbar>
 
                 <!-- Main Content -->
